Modification tipe data for Entity Country and State

+++ src/main/java/com/hendro/ecommerce/entity/Country.java
@@ -16,7 +16,7 @@ public class Country {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id")
-    private short id;
+    private int id;
 
     @Column(name="code")
     private String code;
