+++ src/main/java/com/hendro/ecommerce/entity/State.java
@@ -11,7 +11,7 @@ public class State {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name="id")
-    private short id;
+    private int id;
 
     @Column(name="name")
     private String name;
